+++ lib/model/state/server/receive_session_state.dart
@@ -20,6 +20,7 @@ class ReceiveSessionState with _$ReceiveSessionState {
     required int? startTime,
     required int? endTime,
     required String destinationDirectory,
+    required bool saveToGallery,
 
     // use this to accept / decline the request, empty map == decline
     // FileId -> File Name
@@ -31,5 +32,11 @@ class ReceiveSessionState with _$ReceiveSessionState {
   String? get message {
     final firstFile = files.values.first.file;
     return files.length == 1 && firstFile.fileType == FileType.text ? firstFile.preview : null;
+  }
+
+  /// Returns true if this request contains files having a directory path.
+  /// "Save to gallery" is disabled for such requests.
+  bool get containsDirectories {
+    return files.values.any((f) => f.file.fileName.contains('/'));
   }
 }