+++ lib/model/state/server/server_state.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:localsend_app/model/state/send/web/web_send_state.dart';
 import 'package:localsend_app/model/state/server/receive_session_state.dart';
 
 part 'server_state.freezed.dart';
@@ -13,5 +14,6 @@ class ServerState with _$ServerState {
     required int port,
     required bool https,
     required ReceiveSessionState? session,
+    required WebSendState? webSendState,
   }) = _ServerState;
 }