+++ lib/pages/debug/multicast_debug_page.dart
@@ -4,6 +4,7 @@ import 'package:intl/intl.dart';
 import 'package:localsend_app/provider/multicast_logs_provider.dart';
 import 'package:localsend_app/provider/network/nearby_devices_provider.dart';
 import 'package:localsend_app/widget/copyable_text.dart';
+import 'package:localsend_app/widget/responsive_list_view.dart';
 
 final _dateFormat = DateFormat.Hms();
 
@@ -17,7 +18,7 @@ class MulticastDebugPage extends ConsumerWidget {
       appBar: AppBar(
         title: const Text('Multicast'),
       ),
-      body: ListView(
+      body: ResponsiveListView(
         padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 30),
         children: [
           Row(