+++ lib/pages/tabs/settings_tab.dart
@@ -6,7 +6,7 @@ import 'package:localsend_app/gen/strings.g.dart';
 import 'package:localsend_app/pages/about_page.dart';
 import 'package:localsend_app/pages/changelog_page.dart';
 import 'package:localsend_app/pages/language_page.dart';
-import 'package:localsend_app/provider/network/server_provider.dart';
+import 'package:localsend_app/provider/network/server/server_provider.dart';
 import 'package:localsend_app/provider/settings_provider.dart';
 import 'package:localsend_app/provider/version_provider.dart';
 import 'package:localsend_app/theme.dart';
@@ -228,11 +228,7 @@ class _SettingsTabState extends ConsumerState<SettingsTab> {
                           style: TextButton.styleFrom(foregroundColor: Theme.of(context).colorScheme.onSurface),
                           onPressed: () async {
                             try {
-                              await ref.read(serverProvider.notifier).startServer(
-                                    alias: settings.alias,
-                                    port: settings.port,
-                                    https: settings.https,
-                                  );
+                              await ref.read(serverProvider.notifier).startServerFromSettings();
                             } catch (e) {
                               context.showSnackBar(e.toString());
                             }