+++ assets/CHANGELOG.md
@@ -1,13 +1,16 @@
 ## 1.9.0 (2023-)
 
 - feat: directory share
-- feat: allow deleting history entry when file could not be opened (by @TheGB0077)
+- feat: share via link so users can download files via browser
+- feat: add "delete from history" button when file could not be opened (by @TheGB0077)
 - feat: close message request when copied / opened link
+- feat(android, ios): add "Save to gallery" setting button in file receive options
 - feat(desktop): move troubleshoot out of navigation into send page
 - feat(desktop): save last window position (by @TheGB0077)
 - feat(android): enable edge-to-edge mode
 - feat(android): add monochrome app icons for Android 13 (by @h9419)
 - fix: clear selection after finished message transfer
+- fix(ios): could not scan local network on iOS 14+ (by @TheGB0077)
 - fix(android, ios): fallback asset picker strings to English translation (by @TheGB0077)
 - fix(linux): header bar glitches